refactor(warnings): split glyph pattern helpers
Move pure glyph string-pattern detection out of the page warning emitter so glyphText focuses on warning decisions while the reusable Unicode heuristics stay isolated and easier to review.
